add : room Handler, fix : handler, response, notification directory modify
1. handler, response, notification 폴더 모듈화하여 관리하도록 수정

2. room 생성, 참가 기능 구현
