+++ game/src/classes/managers/base.manager.js
@@ -1,10 +1,9 @@
 class BaseManager {
-    constructor() {
-      if (new.target === BaseManager) {
-        throw new TypeError('Cannot construct MaseManager intstances');
-      }
+  constructor() {
+    if (new.target === BaseManager) {
+      throw new TypeError('Cannot construct BaseManager intstances');
     }
   }
-  
-  export default BaseManager;
-  
+}
+
+export default BaseManager;
