+++ game/src/constants/env.js
@@ -20,6 +20,3 @@ export const REDIS_HOST = process.env.REDIS_HOST;
 export const REDIS_PORT = process.env.REDIS_PORT;
 export const REDIS_PASSWORD = process.env.REDIS_PASSWORD;
 export const SECRET_KEY = process.env.SECRET_KEY;
-
-// 테스트용 토큰
-export const Test_Token = process.env.Test_Token;
