+++ game/src/constants/game.js
@@ -3,3 +3,5 @@ export const MAX_PLAYER = 4;
 export const MAX_PLAYER_HP = 3;
 
 export const MAX_GHOST_NUM = 5;
+
+export const INVITE_CODE_LENGTH = 10;
