+++ game/src/constants/redis.js
@@ -1,0 +1,1 @@
+export const REDIS_USER_SET_KEY = 'USER';
