--- game/src/utils/temp/uuid.temp.js
@@ -1,5 +1,0 @@
-let userNumber = 1;
-
-export const getUUID = () => {
-  return userNumber++;
-};
