+++ package.json
@@ -12,6 +12,7 @@
   "license": "MIT",
   "dependencies": {
     "bcrypt": "^5.1.1",
+    "crypto": "^1.0.1",
     "dotenv": "^16.4.5",
     "ioredis": "^5.4.1",
     "jsonwebtoken": "^9.0.2",
